+++ src/compas_fab/ghpython/artists.py
@@ -9,8 +9,6 @@ from compas.plugins import plugin
 from compas_fab.robots import ReachabilityMap
 
 if compas.RHINO:
-    from compas_ghpython.artists import RobotModelArtist
-
     from .reachability_artist import ReachabilityMapArtist
 
     __all__ = [
@@ -20,11 +18,3 @@ if compas.RHINO:
     @plugin(category="factories", requires=["Rhino"])
     def register_artists():
         Artist.register(ReachabilityMap, ReachabilityMapArtist, context="Grasshopper")
-
-    # deprecated artists (aliased for backwards compat)
-    __all__ += [
-        "RobotArtist",  # deprecated
-        "RobotModelArtist",  # deprecated
-    ]
-
-    RobotArtist = RobotModelArtist